moved deserialize_json and added serialize_json to _util
