+++ tests/core/test_io.py
@@ -10,6 +10,7 @@ from contextlib import nullcontext
 import pytest
 
 from singer_sdk._singerlib import RecordMessage
+from singer_sdk.helpers._util import deserialize_json
 from singer_sdk.io_base import SingerReader, SingerWriter
 
 
@@ -52,9 +53,8 @@ class DummyReader(SingerReader):
     ],
 )
 def test_deserialize(line, expected, exception):
-    reader = DummyReader()
     with exception:
-        assert reader.deserialize_json(line) == expected
+        assert deserialize_json(line) == expected
 
 
 # Benchmark Tests
@@ -104,10 +104,8 @@ def test_bench_deserialize_json(benchmark, bench_encoded_record):
     """Run benchmark for Sink._validator method validate."""
     number_of_runs = 1000
 
-    reader = DummyReader()
-
     def run_deserialize_json():
         for record in itertools.repeat(bench_encoded_record, number_of_runs):
-            reader.deserialize_json(record)
+            deserialize_json(record)
 
     benchmark(run_deserialize_json)
